corrections

+++ src/DynProgPricingAlgo.cpp
@@ -275,7 +275,7 @@ bool DynProgPricingAlgo::findImprovingSolution(InstanceUCP* inst, const DualCost
 
    objvalue = fmin(V_up, V_down) - Dual.Sigma[Site] ;
 
-    if (objvalue <= std::numeric_limits<float>::max()) {
+    if (objvalue <= std::numeric_limits<double>::max()) {
         return true ;
     }
     return false ;
